Server and start of client code
Most of the networking work should be done. The server handles a lot of core game functions like turn order, shuffling cards, and dealing cards. Most of the transmissions that the client can receive from the server have responses set in page.html as in-line JavaScript; however, a few of the transmissions that need to be sent to the server are reliant on UI elements that don't yet exist (such as the way a player will play a card), so they are not currently implemented. I tried to leave plenty of comments for clarification.